🚚 [dockerVM] add dockerVM.Run
Enable calling dockerVM.Run from other repositories
