+++ aws/scenarios/dockerVM/Pulumi.yaml
@@ -1,3 +1,3 @@
-name: docker-vm
+name: dockerVM
 runtime: go
 description: Agent VM with Docker
